Add Docker support to the project
This commit introduces Docker support to the project. It includes a Dockerfile for creating Docker images, a .dockerignore file for specifying which files should not be included in the Docker context, and a compose.yaml file for configuring the Docker setup. Additionally, a new README.Docker.md file has been added to provide instructions on building, running, and deploying Docker applications.